+++ config/uwsm/env-hyprland.d/00-custom.sh
@@ -1,0 +1,1 @@
+HYPRLAND_CONFIG="$HOME/.config/hypr/hyprland.conf"